+++ examples/accelerometer.py
@@ -29,45 +29,24 @@ if len(metawear_devices) < 1:
 else:
     address = metawear_devices[0][0]
 
-c = MetaWearClient(str(address), 'pybluez', debug=False)
+c = MetaWearClient(str(address), 'pybluez', debug=True)
 print("New client created: {0}".format(c))
 
 
 def acc_callback(data):
-    if (data.contents.type_id == DataTypeId.CARTESIAN_FLOAT):
-        data_ptr = cast(data.contents.value, POINTER(CartesianFloat))
-        print("X: {0}, Y: {1}, Z: {2}".format(data_ptr.contents.x,
-                                              data_ptr.contents.y,
-                                              data_ptr.contents.z))
-    else:
-        raise RuntimeError('Incorrect data type id: ' +
-                           str(data.contents.type_id))
+    """Handle a (x,y,z) accelerometer tuple."""
+    print("X: {0}, Y: {1}, Z: {2}".format(*data))
 
 
 print("Write accelerometer settings...")
-libmetawear.mbl_mw_acc_set_odr(c.board, c_float(50.0))
-libmetawear.mbl_mw_acc_set_range(c.board, c_float(2.0))
-#print("Enable logging on board...")
-#libmetawear.mbl_mw_logging_start(c.board)
-print("Write accelerometer config....")
-libmetawear.mbl_mw_acc_write_acceleration_config(c.board)
-print("Subscribing to accelerometer signal...")
+c.accelerometer.set_settings(data_rate=200.0, data_range=2)
+print("Subscribing to accelerometer signal notifications...")
 c.accelerometer.notifications(acc_callback)
-print("Enable acc sampling on board...")
-libmetawear.mbl_mw_acc_enable_acceleration_sampling(c.board)
-print("Start acc on board...")
-libmetawear.mbl_mw_acc_start(c.board)
 
 time.sleep(20.0)
 
 print("Unsubscribe to notification...")
 c.accelerometer.notifications(None)
-print("Stop accelerometer...")
-libmetawear.mbl_mw_acc_stop(c.board)
-print("Disable accelerometer sampling...")
-libmetawear.mbl_mw_acc_disable_acceleration_sampling(c.board)
-#print("Stop logging...")
-#libmetawear.mbl_mw_logging_stop(c.board)
 
 time.sleep(5.0)