+++ examples/led.py
@@ -15,11 +15,8 @@ from __future__ import unicode_literals
 from __future__ import absolute_import
 
 import time
-from ctypes import byref
 
-from pymetawear.client import discover_devices, MetaWearClient, libmetawear
-from pymetawear.mbientlab.metawear.peripheral import Led
-
+from pymetawear.client import discover_devices, MetaWearClient
 
 print("Discovering nearby MetaWear boards...")
 metawear_devices = discover_devices(timeout=2)
@@ -37,10 +34,14 @@ c.led.write_pattern(pattern, 'g')
 c.led.play()
 
 # Same behaviour using libmetawear directly:
-# pattern = Led.Pattern(repeat_count=10)
-# libmetawear.mbl_mw_led_load_preset_pattern(byref(pattern), Led.PRESET_BLINK)
-# libmetawear.mbl_mw_led_write_pattern(c.board, byref(pattern), Led.COLOR_GREEN)
-# libmetawear.mbl_mw_led_play(c.board)
+#from pymetawear import libmetawear
+#from ctypes import byref
+#from pymetawear.mbientlab.metawear.peripheral import Led
+#
+#pattern = Led.Pattern(repeat_count=10)
+#libmetawear.mbl_mw_led_load_preset_pattern(byref(pattern), Led.PRESET_BLINK)
+#libmetawear.mbl_mw_led_write_pattern(c.board, byref(pattern), Led.COLOR_GREEN)
+#libmetawear.mbl_mw_led_play(c.board)
 
 time.sleep(5.0)